requirements.txt
python-twitch-stream
opencv-python
tensorflow-gpu
scikit-image
keras